requirements.txt
selenium==4.19.0
python-dotenv==1.0.1
webdriver-manager==4.0.0
pytest==8.1.1
pytest-xdist==3.5.0
pytest-html==4.1.1
